perf: compatibility cleanup - use std::exp2/std::pow with explicit casts and include &lt;cmath&gt;

+++ Source/CS01Synth/VCAProcessor.cpp
@@ -1,4 +1,5 @@
 #include "VCAProcessor.h"
+#include <cmath>
 
 //==============================================================================
 VCAProcessor::VCAProcessor(juce::AudioProcessorValueTreeState& apvts)
@@ -113,7 +114,7 @@ void VCAProcessor::processBlock(juce::AudioBuffer<float>& buffer, juce::MidiBuff
 // IG02600 VCA chip emulation
 float VCAProcessor::processVCA(float input, float controlVoltage, float volumeParam) {
     // Apply logarithmic volume control characteristic (PVR5)
-    float volume = std::pow(volumeParam, 2.5f);  // Use std::pow (float overload) for portability
+    float volume = static_cast<float>(std::pow(static_cast<double>(volumeParam), 2.5));  // Use std::pow with explicit casts for portability
 
     // Calculate gain based on control voltage and volume
     float gain = controlVoltage * volume;
